requirements.txt
pyTelegramBotAPI==4.14.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
schedule==1.2.0
urllib3==2.0.7
certifi==2023.7.22